{"schema":"decodex/commit/1","summary":"Modularize state run-control and history boundaries","authority":"manual"}
