+++ apps/decodex/src/agent/app_server/turn_loop/wait.rs
@@ -1,46 +1,27 @@
-use std::time::{Duration, Instant};
+mod pending;
+mod request;
+mod timeout;
+mod wire;
 
-use color_eyre::eyre::Report;
+pub(in crate::agent::app_server) use self::{
+	pending::flush_pending_messages, timeout::is_app_server_output_timeout,
+};
+
+use std::time::Instant;
 
 use crate::{
 	agent::{
 		app_server::{
-			AppServerClient, activity,
-			constants::RUN_CONTROL_POLL_INTERVAL,
-			lane_control,
-			runtime_types::{
-				AppServerRunRequest, RequestDispatchContext, RequestWaitPhase, RunRecorder,
-			},
-			server_requests, transport,
+			AppServerClient, activity, lane_control,
+			runtime_types::{AppServerRunRequest, RunRecorder},
+			server_requests,
 			turn_failure::AppServerTurnFailure,
 			turn_loop::{RunOutcome, completion, messages},
 		},
-		codex_accounts::CodexAccountProvider,
-		json_rpc::{AppServerOutputTimeout, JsonRpcMessage, JsonRpcRequest, WireMessage},
-		tracker_tool_bridge::DynamicToolHandler,
+		json_rpc::JsonRpcMessage,
 	},
-	prelude::{Result, eyre},
+	prelude::Result,
 };
-
-pub(in crate::agent::app_server) fn flush_pending_messages(
-	client: &mut AppServerClient,
-	recorder: &mut RunRecorder<'_>,
-	target_thread_id: Option<&str>,
-) -> Result<()> {
-	for message in client.drain_pending() {
-		if messages::targets_thread(&message, target_thread_id) {
-			recorder.record(messages::message_type(&message), &message.raw)?;
-
-			server_requests::apply_protocol_message_side_effects(recorder, &message)?;
-		}
-	}
-
-	Ok(())
-}
-
-pub(in crate::agent::app_server) fn is_app_server_output_timeout(error: &Report) -> bool {
-	error.downcast_ref::<AppServerOutputTimeout>().is_some()
-}
 
 pub(super) fn wait_for_turn_completion(
 	client: &mut AppServerClient,
@@ -75,7 +56,7 @@ pub(super) fn wait_for_turn_completion(
 			Some(&recorder.protocol_activity.summary),
 			request.timeout,
 		);
-		let Some(wire_message) = next_turn_wire_message(
+		let Some(wire_message) = wire::next_turn_wire_message(
 			client,
 			last_activity_at,
 			idle_timeout,
@@ -119,7 +100,7 @@ pub(super) fn wait_for_turn_completion(
 					return Ok(outcome);
 				}
 			},
-			JsonRpcMessage::Request(server_request) => handle_turn_execution_request(
+			JsonRpcMessage::Request(server_request) => request::handle_turn_execution_request(
 				client,
 				recorder,
 				server_request,
@@ -128,7 +109,7 @@ pub(super) fn wait_for_turn_completion(
 				request.dynamic_tool_handler,
 				request.codex_account_provider,
 			)?,
-			JsonRpcMessage::Response(_) => ignore_orphan_turn_json_rpc_response(),
+			JsonRpcMessage::Response(_) => request::ignore_orphan_turn_json_rpc_response(),
 			JsonRpcMessage::Error(error) => {
 				latest_turn_failure = Some(completion::turn_failure_from_json_rpc_error_response(
 					target_thread_id,
@@ -137,101 +118,5 @@ pub(super) fn wait_for_turn_completion(
 				));
 			},
 		}
-	}
-}
-
-fn next_turn_wire_message(
-	client: &mut AppServerClient,
-	last_activity_at: Instant,
-	timeout: Duration,
-	target_thread_id: &str,
-	target_turn_id: &str,
-	latest_turn_failure: Option<&AppServerTurnFailure>,
-	control_enabled: bool,
-) -> Result<Option<WireMessage>> {
-	let now = Instant::now();
-	let wait_timeout =
-		messages::remaining_idle_budget(last_activity_at, now, timeout).ok_or_else(|| {
-			turn_wait_timeout_error(target_thread_id, target_turn_id, latest_turn_failure.cloned())
-		})?;
-	let recv_timeout =
-		if control_enabled { wait_timeout.min(RUN_CONTROL_POLL_INTERVAL) } else { wait_timeout };
-
-	match recv_turn_wire_message(client, recv_timeout, latest_turn_failure) {
-		Ok(wire_message) => Ok(Some(wire_message)),
-		Err(error)
-			if control_enabled
-				&& recv_timeout < wait_timeout
-				&& is_app_server_output_timeout(&error) =>
-			Ok(None),
-		Err(error) => Err(error),
-	}
-}
-
-fn handle_turn_execution_request(
-	client: &mut AppServerClient,
-	recorder: &mut RunRecorder<'_>,
-	request: &JsonRpcRequest,
-	target_thread_id: &str,
-	target_turn_id: &str,
-	dynamic_tool_handler: Option<&dyn DynamicToolHandler>,
-	codex_account_provider: Option<&dyn CodexAccountProvider>,
-) -> Result<()> {
-	server_requests::handle_server_request_during_turn_execution(
-		client,
-		recorder,
-		request,
-		RequestDispatchContext::new(
-			RequestWaitPhase::TurnExecution,
-			dynamic_tool_handler,
-			codex_account_provider,
-			Some(target_thread_id),
-			Some(target_turn_id),
-		),
-	)
-}
-
-fn ignore_orphan_turn_json_rpc_response() {
-	tracing::debug!(
-		"Recorded and ignored orphan app-server JSON-RPC response while waiting for turn completion."
-	);
-}
-
-fn turn_wait_timeout_error(
-	target_thread_id: &str,
-	target_turn_id: &str,
-	latest_turn_failure: Option<AppServerTurnFailure>,
-) -> Report {
-	let message = format!(
-		"Timed out while waiting for turn `{target_turn_id}` on thread `{target_thread_id}`."
-	);
-
-	if let Some(failure) = latest_turn_failure {
-		return Report::new(failure).wrap_err(message);
-	}
-
-	eyre::eyre!(message)
-}
-
-fn recv_turn_wire_message(
-	client: &mut AppServerClient,
-	wait_timeout: Duration,
-	latest_turn_failure: Option<&AppServerTurnFailure>,
-) -> Result<WireMessage> {
-	match transport::annotate_transport_failure_phase(
-		client.recv(Some(wait_timeout)),
-		RequestWaitPhase::TurnExecution,
-	) {
-		Ok(wire_message) => Ok(wire_message),
-		Err(error) => {
-			if error.downcast_ref::<AppServerOutputTimeout>().is_some()
-				&& let Some(failure) = latest_turn_failure
-			{
-				return Err(Report::new(failure.clone())
-					.wrap_err("Timed out while waiting for additional app-server output."));
-			}
-
-			Err(error)
-		},
 	}
 }
